api_yamdb/requirements.txt
aiohttp==3.8.3
aiomoex==1.2.2
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.11.6
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dill==0.3.5.1
Django==2.2.16
django-debug-toolbar==3.2.4
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.0.4
Faker==12.0.1
frozenlist==1.3.1
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
investments==0.1.25
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mixer==7.1.2
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
Pillow==8.3.1
platformdirs==2.5.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pyaes==1.6.1
pycparser==2.21
PyJWT==2.1.0
pylint==2.13.9
pyparsing==3.0.9
Pyrogram==2.0.30
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv
python3-openid==3.2.0
pytz==2022.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tabulate==0.8.10
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
